requirements.txt
absl-py==0.7.0
astor==0.7.1
Django==2.1.7
django-pyodbc-azure==2.1.0.0
gast==0.2.2
grpcio==1.18.0
h5py==2.9.0
imageio==2.5.0
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
Markdown==3.0.1
mock==3.0.5
numpy==1.16.1
opencv-python==4.0.0.21
Pillow==5.4.1
protobuf==3.6.1
pyodbc==4.0.26
pytz==2018.9
scikit-learn==0.20.2
scipy==1.2.1
six==1.12.0
sklearn==0.0
tensorboard==1.13.1
tensorflow==1.13.1
tensorflow-estimator==1.13.0
termcolor==1.1.0
Werkzeug==0.14.1
